+++ src/main/kotlin/com/github/strindberg/emacsj/universal/UniversalArgumentDelegate.kt
@@ -1,5 +1,6 @@
 package com.github.strindberg.emacsj.universal
 
+import com.github.strindberg.emacsj.line.ACTION_TRANSPOSE_LINES
 import com.github.strindberg.emacsj.mark.ACTION_POP_MARK
 import com.github.strindberg.emacsj.mark.ACTION_PUSH_MARK
 import com.github.strindberg.emacsj.paste.ACTION_PASTE
@@ -60,6 +61,8 @@ class UniversalArgumentDelegate(val editor: Editor, private var numeric: Int?) {
             ACTION_ZAP_FORWARD_UP_TO,
             ACTION_ZAP_BACKWARD_TO,
             ACTION_ZAP_BACKWARD_UP_TO,
+            ACTION_DELETE_SPACE,
+            ACTION_TRANSPOSE_LINES,
         )
 
         public fun registerSingleAction(actionId: String) {