+++ CHANGELOG.md
@@ -5,9 +5,7 @@
 ## [Unreleased]
 
 Note: previous versions of this plugin modified the base IntelliJ keymap, adding the plugin's bindings to all keymaps. This was incorrect
-and has been removed in this release. If this causes unforeseen consequences, please file an issue in the GitHub project. The plugin
-still adds its commands to the Emacs keymap, which is also questionable, but I have kept it so for this release as I don't know if removing
-it would inconvenience users relying on this behavior. File an issue if you think this is a problem.
+and has been removed in this release. If this causes unforeseen consequences, please file an issue on GitHub.
 
 ### Added