+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+## [1.5.2] - 2025-12-30
+
 Note: previous versions of this plugin modified the base IntelliJ keymap, adding the plugin's bindings to all keymaps. This was incorrect
 and has been removed in this release. If this causes unforeseen consequences, please file an issue on GitHub.
 
@@ -226,7 +228,8 @@ independent commands if you for some reason prefer not to use *Universal Argumen
 
 - Initial release.
 
-[Unreleased]: https://github.com/strindberg/emacsj/compare/v1.5.1...HEAD
+[Unreleased]: https://github.com/strindberg/emacsj/compare/v1.5.2...HEAD
+[1.5.2]: https://github.com/strindberg/emacsj/compare/v1.5.1...v1.5.2
 [1.5.1]: https://github.com/strindberg/emacsj/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/strindberg/emacsj/compare/v1.4.3...v1.5.0
 [1.4.3]: https://github.com/strindberg/emacsj/compare/v1.4.2...v1.4.3